backend/pyproject.toml
[tool.poetry]
name = "sig-backend"
version = "0.0.1"
description = "The backend API for Sig project."
authors = ["EshaanAgg <eshaanagg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pysigma-backend-azure = {git = "https://github.com/sifex/pySigma-backend-azure.git", rev = "9c35b78cd4d4891e1b46ccf21ed555ffb10b96e6"}
pysigma-backend-loki = {git = "https://github.com/grafana/pySigma-backend-loki.git", rev = "452aa0d8bb096bbabdca5a83a884c45662dec666"}
pysigma-backend-microsoft365defender = {git = "https://github.com/AttackIQ/pySigma-backend-microsoft365defender.git", rev = "731db4b6c7cbab8898973b4350d93268e135c495"}
pysigma-backend-stix = {git = "https://github.com/barvhaim/pySigma-backend-stix", rev = "0d7c05e187249c26f5abb99f63bf839c78705d1e"}
flask = "3.0.0"
sigma-cli = "0.7.6"
pysigma = "0.9.11"
pysigma-backend-carbonblack = "0.1.4"
pysigma-backend-cortexxdr = "0.1.1"
pysigma-backend-elasticsearch = "1.0.5"
pysigma-backend-opensearch = "1.0.0"
pysigma-backend-qradar = "0.3.3"
pysigma-backend-qradar-aql = "0.2.3"
pysigma-backend-sentinelone = "0.1.2"
pysigma-backend-sentinelone-pq = "0.1.1"
pysigma-backend-splunk = "1.0.2"
pysigma-backend-insightidr = "0.2.1"
pysigma-pipeline-crowdstrike = "1.0.0"
pysigma-pipeline-sysmon = "1.0.2"
pysigma-pipeline-windows = "1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
